Rename files to reflect that they are not just reg num related
